moved range proofs to a separate file
