+++ README.md
@@ -1,4 +1,19 @@
-# ristretto-bulletproofs
+# Ristretto Bulletproofs
+
 Implementing [bulletproofs](https://crypto.stanford.edu/bulletproofs/) using [ristretto](https://github.com/dalek-cryptography/ed25519-dalek).
 
 Step 1 of a larger proposed proposed plan, detailed [here](https://github.com/chain/research/issues/7).
+
+## Development
+
+Run tests:
+
+```
+cargo test
+```
+
+Run benchmarks:
+
+```
+cargo bench --features="bench"
+```
